cardiac-ai/counterfactual-explanation-microservice/requirements.txt
numpy==1.19.5
scipy==1.4.1
jax==0.2.17
jaxlib==0.1.69
cvxpy==1.1.0
scikit-learn==0.24.2
sklearn-lvq==1.1.1
tensorflow==2.5.2
torch==1.7.1
sphinx==2.1.2
sphinx-rtd-theme==0.4.3
pytest==5.0.1
Django==4.0.2
djangorestframework
django-filter
markdown
django-cors-headers==3.5.0
pika==1.1.0